cleaning up session
