+++ sources/jobs/append_cmd_arr.c
