+++ sources/jobs/append_pid_arr.c
