+++ sources/jobs/bg_node_delete.c
