+++ sources/jobs/close_all_bg_processes.c
