+++ sources/jobs/dup_dbl_ptr.c
