+++ sources/jobs/triple_ptr_len.c
